型変換の誤り修正/浮動小数点種別判定周り,copysignのfe_holder対応追加/

+++ include/tunum/math/ceil.hpp
@@ -79,7 +79,7 @@ namespace tunum::_math_impl
     template <class T>
     requires (is_fe_holder_v<T>)
     inline constexpr auto ceil(const T& x) noexcept
-    { return fe_ceil(double(x)); }
+    { return fe_ceil(x); }
 
     // --------------------------------------
     // cpo
